pyproject.toml
[tool.poetry]
name = "referral-api"
version = "0.1.0"
description = ""
authors = ["Danil Galeev <galoffdan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
uwsgi = "^2.0.27"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
setuptools = "^75.1.0"
djangorestframework-simplejwt = "^5.3.1"
django-oauth-toolkit = "^3.0.1"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
ipython = "^8.28.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120